Created the Main Menu
I created a Main Menu with a start and quit button. The start button opens the level prototype map and the quit button exits the game. I also added some bluprint logic to the level prototype. It changes the input to game only so there are no issues after opening the level prototype from the main menu. I also moved a few static objects in the level prototype.